rename database start file
